vLLMClient の廃止 (#46)
* Add Ruff configuration

* Make environment variables configurable

* Remove vLLMClient

* Update README

* Apply suggestions from code review


* Rename client.local to client.base

* Update README (description on vLLM client)

* Update README (description on vLLM client)

---------
