+++ src/llm_jp_judge/client/__init__.py
@@ -1,4 +1,3 @@
-from .local import vLLMClient
 from .remote import AzureOpenAI, BedrockAnthropic, OpenAI
 
 
@@ -9,6 +8,4 @@ def load_client(name="azure", **kwargs):
         return AzureOpenAI(**kwargs)
     elif name == "bedrock":
         return BedrockAnthropic(**kwargs)
-    elif name == "vllm":
-        return vLLMClient(**kwargs)
     raise ValueError(f"Invalid client name: {name}")
